requirements.txt
certifi==2018.11.29
Click==7.0
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-extensions==2.1.5
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
flasgger==0.9.2 
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.17.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
pep8==1.7.1
pipenv==2018.11.26
psycopg2==2.7.7
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.9
PyYAML==3.13
six==1.12.0
virtualenv==16.3.0
virtualenv-clone==0.5.1
Werkzeug==0.14.1
wrapt==1.11.1
